fix gracefull exit
